PropertiesChainBuilder
