+++ com.sap.cloud.lm.sl.mta/src/main/java/com/sap/cloud/lm/sl/mta/builders/v2/PropertiesChainBuilder.java
@@ -17,21 +17,27 @@ import com.sap.cloud.lm.sl.mta.model.v2.RequiredDependency;
 import com.sap.cloud.lm.sl.mta.model.v2.Resource;
 import com.sap.cloud.lm.sl.mta.util.PropertiesUtil;
 
-public class PropertiesChainBuilder extends com.sap.cloud.lm.sl.mta.builders.v1.PropertiesChainBuilder {
-
-    public PropertiesChainBuilder(DeploymentDescriptor descriptor, Platform platform) {
-        super(descriptor, platform, new DescriptorHandler());
-    }
-
-    public PropertiesChainBuilder(DeploymentDescriptor descriptor) {
-        super(descriptor, null, new DescriptorHandler());
-    }
+public class PropertiesChainBuilder {
+    
+    protected final DeploymentDescriptor descriptor;
+    protected final Platform platform;
+    protected final DescriptorHandler handler;
 
     public PropertiesChainBuilder(DeploymentDescriptor descriptor, Platform platform, DescriptorHandler handler) {
-        super(descriptor, platform, handler);
+        this.descriptor = descriptor;
+        this.platform = platform;
+        this.handler = handler;
+    }
+    
+    public PropertiesChainBuilder(DeploymentDescriptor descriptor, Platform platform) {
+        this(descriptor, platform, new DescriptorHandler());
+    }
+    
+    public PropertiesChainBuilder(DeploymentDescriptor descriptor) {
+
+        this(descriptor, null, new DescriptorHandler());
     }
 
-    @Override
     public List<Map<String, Object>> buildModuleChain(String moduleName) {
         Module module = (Module) handler.findModule(descriptor, moduleName);
         if (module == null) {
@@ -41,8 +47,15 @@ public class PropertiesChainBuilder extends com.sap.cloud.lm.sl.mta.builders.v1.
         ModuleType moduleType = (ModuleType) getModuleType(module);
         return getPropertiesList(dependencies, module, moduleType);
     }
+    
+    
+    protected ModuleType getModuleType(Module module) {
+        if (platform == null) {
+            return null;
+        }
+        return handler.findModuleType(platform, module.getType());
+    }
 
-    @Override
     public List<Map<String, Object>> buildModuleChainWithoutDependencies(String moduleName) {
         Module module = (Module) handler.findModule(descriptor, moduleName);
         if (module == null) {
@@ -51,19 +64,17 @@ public class PropertiesChainBuilder extends com.sap.cloud.lm.sl.mta.builders.v1.
         ModuleType moduleType = (ModuleType) getModuleType(module);
         return PropertiesUtil.getPropertiesList(module, moduleType);
     }
+    
+    public List<Map<String, Object>> buildResourceTypeChain(String resourceName) {
+        throw new UnsupportedOperationException();
+    }
 
-    @Override
     public List<Map<String, Object>> buildResourceChain(String resourceName) {
         Resource resource = (Resource) handler.findResource(descriptor, resourceName);
         if (resource == null) {
             return Collections.emptyList();
         }
         return PropertiesUtil.getPropertiesList(resource);
-    }
-
-    @Override
-    public List<Map<String, Object>> buildResourceTypeChain(String resourceName) {
-        throw new UnsupportedOperationException();
     }
 
     protected static List<Map<String, Object>> getPropertiesList(List<RequiredDependency> dependencies, Module module,
